datacollector/requirements.txt
Scrapy==2.11.2
py-tlsh==4.7.2
tlsh==0.2.0
pymongo==3.12.0
psutil==6.0.0
fastapi==0.111.1
uvicorn==0.30.1
PyYAML==6.0.1
mongomock==4.1.2
numpy==1.24.3
python-dotenv==1.0.1